GPU timestamp queries and block-based profile system

+++ src/sumire/core/profiling/profiling_block.hpp
@@ -6,7 +6,7 @@ namespace sumire {
 
     struct ProfilingBlock {
         uint32_t idx;
-        float ms = 0.0f;
+        double ms = 0.0;
     };
 
 }
